+++ packages/core/tests/Aesthetic.test.tsx
@@ -27,7 +27,6 @@ describe('Aesthetic', () => {
         cxPropName: 'cx',
         extendable: false,
         passThemeProp: false,
-        pure: true,
         rtl: false,
         stylesPropName: 'styleSheet',
         theme: 'default',
@@ -463,14 +462,6 @@ describe('Aesthetic', () => {
   });
 
   describe('registerStyleSheet()', () => {
-    it('errors if styles have been set', () => {
-      instance.styles.foo = () => ({});
-
-      expect(() => {
-        instance.registerStyleSheet('foo', () => ({}));
-      }).toThrowErrorMatchingSnapshot();
-    });
-
     it('errors if styles try to extend from itself', () => {
       instance.styles.foo = () => ({});