+++ packages/react/tests/withStylesFactory.test.tsx
@@ -13,9 +13,9 @@ describe('withStylesFactory()', () => {
 
   beforeEach(() => {
     aesthetic = new TestAesthetic();
-    withStyles = withStylesFactory(aesthetic);
-
     registerTestTheme(aesthetic);
+
+    withStyles = withStylesFactory(aesthetic);
   });
 
   function BaseComponent() {
@@ -40,10 +40,7 @@ describe('withStylesFactory()', () => {
     return shallow(element, {
       // @ts-ignore Not yet typed
       wrappingComponent: WrappingComponent,
-    })
-      .dive()
-      .dive()
-      .dive();
+    });
   }
 
   it('returns an HOC component', () => {
@@ -94,6 +91,8 @@ describe('withStylesFactory()', () => {
       },
     });
     const Wrapped = withStyles(styles)(BaseComponent);
+
+    shallowDeep(<Wrapped />);
 
     expect(aesthetic.styles[Wrapped.styleName]).toBe(styles);
   });
@@ -170,16 +169,13 @@ describe('withStylesFactory()', () => {
   it('creates a style sheet', () => {
     const spy = jest.spyOn(aesthetic, 'createStyleSheet');
     const Wrapped = withStyles(() => TEST_STATEMENT)(StyledComponent);
-    const wrapper = shallowDeep(<Wrapped foo="abc" />);
+
+    shallowDeep(<Wrapped foo="abc" />);
 
     expect(spy).toHaveBeenCalledWith(Wrapped.styleName, {
       dir: 'ltr',
       name: Wrapped.styleName,
-      theme: 'light',
-    });
-    expect(wrapper.state('styles')).toEqual({
-      header: {},
-      footer: {},
+      theme: '',
     });
   });
 
@@ -306,7 +302,7 @@ describe('withStylesFactory()', () => {
     expect(createSpy).toHaveBeenCalledWith(Wrapped.styleName, {
       dir: 'rtl',
       name: Wrapped.styleName,
-      theme: 'default',
+      theme: '',
     });
 
     act(() => {
@@ -321,7 +317,7 @@ describe('withStylesFactory()', () => {
     expect(createSpy).toHaveBeenCalledWith(Wrapped.styleName, {
       dir: 'ltr',
       name: Wrapped.styleName,
-      theme: 'default',
+      theme: '',
     });
   });
 
@@ -384,12 +380,12 @@ describe('withStylesFactory()', () => {
       expect(createSpy).toHaveBeenCalledWith(Wrapped.styleName, {
         dir: 'rtl',
         name: Wrapped.styleName,
-        theme: 'default',
+        theme: '',
       });
       expect(transformSpy).toHaveBeenCalledWith([{}, {}], {
         dir: 'rtl',
         name: Wrapped.styleName,
-        theme: 'default',
+        theme: '',
       });
     });
 
@@ -411,12 +407,12 @@ describe('withStylesFactory()', () => {
       expect(createSpy).toHaveBeenCalledWith(Wrapped.styleName, {
         dir: 'rtl',
         name: Wrapped.styleName,
-        theme: 'default',
+        theme: '',
       });
       expect(transformSpy).toHaveBeenCalledWith([{}, {}], {
         dir: 'rtl',
         name: Wrapped.styleName,
-        theme: 'default',
+        theme: '',
       });
     });
   });