+++ packages/core/CHANGELOG.md
@@ -11,13 +11,13 @@ slightly change.
   [aesthetic-react](https://github.com/milesj/aesthetic/tree/master/packages/react) package.
 - Helper functions have moved to the new
   [aesthetic-utils](https://github.com/milesj/aesthetic/tree/master/packages/utils) package.
-- Updated `Aesthetic#getTheme` to require a name argument. Will no longer default to active theme.
 - Updated `Aesthetic#transformStyles` to require all styles as an array in the 1st argument, instead
   of being spread across all arguments.
 - Updated `Aesthetic#transformToClassName` to only be passed the parsed blocks and not the native
   adapter blocks.
 - Renamed `Aesthetic#setStyleSheet` to `registerStyleSheet`.
 - Renamed `Aesthetic#processStyleSheet` to `parseStyleSheet`.
+- Removed the `pure` option (since class components are no longer used).
 
 #### 🚀 Updates