+++ packages/react/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Updated `react` requirement to v16.6.
 - Updated `withStyles` HOC to receive the CSS transformer function as a `cx` prop.
+- Updated `withStyles` and `withTheme` HOCs to internally use function components instead of class
+  components.
 - **[TS]** Renamed the `WithStylesProps` interface to `WithStylesWrappedProps`.
 - **[TS]** Renamed the `WithThemeProps` interface to `WithThemeWrappedProps`.